sample_project/requirements.txt
django-forms-bootstrap==3.1.0
freeze==1.0.10
olefile==0.44
Pillow==4.0.0
pkg-resources==0.0.0
six==1.11.0
